move manifest
